// src/api/__generated__/story/model/publishShortVideoRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface PublishShortVideoRequest {
  uploadSessionId: number;
  videoObjectKey: string;
  coverObjectKey: string;
  /**
   * 选填,为空时存空字符串
   * @minLength 0
   * @maxLength 200
   */
  title?: string;
  /**
   * 必填
   * @minLength 0
   * @maxLength 1000
   */
  description: string;
  durationSec: number;
  width: number;
  height: number;
}